tests: make another pass at cleaning up flaky tests (#1537)
* tests: make schema tests more deterministic

* tests: clean up sample subscription tests better

* 🦉 Updates from OwlBot post-processor

See https://github.com/googleapis/repo-automation-bots/blob/main/packages/owl-bot/README.md

* 🦉 Updates from OwlBot post-processor

See https://github.com/googleapis/repo-automation-bots/blob/main/packages/owl-bot/README.md

+++ handwritten/pubsub/samples/system-test/schema.test.ts
@@ -47,19 +47,17 @@ describe('schema', () => {
   // Schemas have a delay between deletion and actual deletion, so
   // we have to make unique schema names. It's simplest to do it for topics
   // and subscriptions too.
-  let curTopic = 1;
-  function getTopicId(): string {
-    return topicIdStem + curTopic++;
+  let curId = 1;
+  function getTopicId(testName: string): string {
+    return `${topicIdStem}${testName}${curId++}`;
   }
 
-  let curSub = 1;
-  function getSubId(): string {
-    return subscriptionIdStem + curSub++;
+  function getSubId(testName: string): string {
+    return `${subscriptionIdStem}${testName}${curId++}`;
   }
 
-  let curSchema = 1;
-  function getSchemaId(): string {
-    return schemaIdStem + curSchema++;
+  function getSchemaId(testName: string): string {
+    return `${schemaIdStem}${testName}${curId++}`;
   }
 
   function fullSchemaName(base: string) {
@@ -68,46 +66,59 @@ describe('schema', () => {
 
   const commandFor = (action: string) => `node ${action}.js`;
 
-  async function cleanAllSchemas() {
+  async function cleanAllSchemas(prefix: string) {
     const schemas = [];
     for await (const s of pubsub.listSchemas()) {
-      schemas.push(pubsub.schema(s.name!).delete());
+      if (s.name?.startsWith(prefix)) {
+        schemas.push(pubsub.schema(s.name!).delete());
+      }
     }
     await Promise.all(schemas);
   }
 
-  async function cleanAllTopics() {
+  async function cleanAllTopics(prefix: string) {
     const [topics] = await pubsub.getTopics();
     await Promise.all(
-      topics.filter(x => x.name.endsWith(runId)).map(x => x.delete())
+      topics.filter(x => x.name.startsWith(prefix)).map(x => x.delete())
     );
   }
 
-  async function cleanAllSubs() {
+  async function cleanAllSubs(prefix: string) {
     const [subscriptions] = await pubsub.getSubscriptions();
     await Promise.all(
-      subscriptions.filter(x => x.name.endsWith(runId)).map(x => x.delete())
+      subscriptions.filter(x => x.name.startsWith(prefix)).map(x => x.delete())
     );
   }
 
+  // This is to make sure we didn't miss something from per-test cleanup.
   after(async () => {
-    await cleanAllSubs();
-    await cleanAllTopics();
-    await cleanAllSchemas();
+    await cleanAllSubs(subscriptionIdStem);
+    await cleanAllTopics(topicIdStem);
+    await cleanAllSchemas(schemaIdStem);
   });
+
+  // Each test should call this to clean its own resources.
+  async function cleanMine(testName: string) {
+    await cleanAllSubs(`${subscriptionIdStem}${testName}`);
+    await cleanAllTopics(`${topicIdStem}${testName}`);
+    await cleanAllSchemas(`${schemaIdStem}${testName}`);
+  }
 
   function fixturePath(fixture: string): string {
     return path.join(__dirname, '..', '..', 'system-test', 'fixtures', fixture);
   }
 
-  async function createSchema(type: 'avro' | 'proto'): Promise<Schema> {
+  async function createSchema(
+    testName: string,
+    type: 'avro' | 'proto'
+  ): Promise<Schema> {
     const suffix = type === 'avro' ? 'avsc' : 'proto';
     const encoding =
       type === 'avro' ? SchemaTypes.Avro : SchemaTypes.ProtocolBuffer;
     const def = (
       await fs.readFile(fixturePath(`provinces.${suffix}`))
     ).toString();
-    const schemaId = getSchemaId();
+    const schemaId = getSchemaId(testName);
     const schema = await pubsub.createSchema(schemaId, encoding, def);
     assert.ok(schema);
 
@@ -115,10 +126,11 @@ describe('schema', () => {
   }
 
   async function createTopicWithSchema(
+    testName: string,
     schemaName: string,
     encodingType: SchemaEncoding
   ): Promise<Topic> {
-    const topicId = getTopicId();
+    const topicId = getTopicId(testName);
     const [topic] = await pubsub.createTopic({
       name: topicId,
       schemaSettings: {
@@ -131,8 +143,11 @@ describe('schema', () => {
     return topic;
   }
 
-  async function createSub(topicName: string): Promise<Subscription> {
-    const subId = getSubId();
+  async function createSub(
+    testName: string,
+    topicName: string
+  ): Promise<Subscription> {
+    const subId = getSubId(testName);
     const [sub] = await pubsub.createSubscription(topicName, subId);
     assert.ok(sub);
 
@@ -140,173 +155,231 @@ describe('schema', () => {
   }
 
   it('should create an avro schema', async () => {
-    const schemaId = getSchemaId();
-    const output = execSync(
-      `${commandFor('createAvroSchema')} ${schemaId} ${fixturePath(
-        'provinces.avsc'
-      )}`
-    );
-    assert.include(output, schemaId);
-    assert.include(output, 'created.');
+    const id = 'create-avro';
+    try {
+      const schemaId = getSchemaId(id);
+      const output = execSync(
+        `${commandFor('createAvroSchema')} ${schemaId} ${fixturePath(
+          'provinces.avsc'
+        )}`
+      );
+      assert.include(output, schemaId);
+      assert.include(output, 'created.');
 
-    let found = false;
-    for await (const s of pubsub.listSchemas()) {
-      if (s.name && s.name.indexOf(schemaId) >= 0) {
-        found = true;
-        break;
+      let found = false;
+      for await (const s of pubsub.listSchemas()) {
+        if (s.name && s.name.indexOf(schemaId) >= 0) {
+          found = true;
+          break;
+        }
       }
-    }
 
-    assert.ok(found, 'created schema was not found');
+      assert.ok(found, 'created schema was not found');
+    } finally {
+      await cleanMine(id);
+    }
   });
 
   it('should create a proto schema', async () => {
-    const schemaId = getSchemaId();
-    const output = execSync(
-      `${commandFor('createProtoSchema')} ${schemaId} ${fixturePath(
-        'provinces.proto'
-      )}`
-    );
-    assert.include(output, schemaId);
-    assert.include(output, 'created.');
+    const id = 'create-proto';
+    try {
+      const schemaId = getSchemaId(id);
+      const output = execSync(
+        `${commandFor('createProtoSchema')} ${schemaId} ${fixturePath(
+          'provinces.proto'
+        )}`
+      );
+      assert.include(output, schemaId);
+      assert.include(output, 'created.');
 
-    let found = false;
-    for await (const s of pubsub.listSchemas()) {
-      if (s.name && s.name.indexOf(schemaId) >= 0) {
-        found = true;
-        break;
+      let found = false;
+      for await (const s of pubsub.listSchemas()) {
+        if (s.name && s.name.indexOf(schemaId) >= 0) {
+          found = true;
+          break;
+        }
       }
-    }
 
-    assert.ok(found, 'created schema was not found');
+      assert.ok(found, 'created schema was not found');
+    } finally {
+      await cleanMine(id);
+    }
   });
 
   it('should create a topic with a schema', async () => {
-    const schema = await createSchema('proto');
-    const topicId = getTopicId();
-    const output = execSync(
-      `${commandFor('createTopicWithSchema')} ${topicId} ${schema.id} BINARY`
-    );
-    assert.include(output, topicId);
-    assert.include(output, schema.id);
-    assert.include(output, 'created with');
+    const id = 'create-topic';
+    try {
+      const schema = await createSchema('topic-schema', 'proto');
+      const topicId = getTopicId(id);
+      const output = execSync(
+        `${commandFor('createTopicWithSchema')} ${topicId} ${schema.id} BINARY`
+      );
+      assert.include(output, topicId);
+      assert.include(output, schema.id);
+      assert.include(output, 'created with');
 
-    const [topic] = await pubsub.topic(topicId).get();
-    assert.include(topic.metadata?.schemaSettings?.schema, schema.id);
+      const [topic] = await pubsub.topic(topicId).get();
+      assert.include(topic.metadata?.schemaSettings?.schema, schema.id);
+    } finally {
+      await cleanMine(id);
+    }
   });
 
   it('should delete a schema', async () => {
-    const schema = await createSchema('proto');
+    const id = 'delete-schema';
+    try {
+      const schema = await createSchema(id, 'proto');
 
-    const output = execSync(`${commandFor('deleteSchema')} ${schema.id}`);
-    assert.include(output, schema.id);
-    assert.include(output, 'deleted.');
+      const output = execSync(`${commandFor('deleteSchema')} ${schema.id}`);
+      assert.include(output, schema.id);
+      assert.include(output, 'deleted.');
 
-    // Because of caching delays, this can't be deterministic without a big wait.
-    /*try {
-      const got = await pubsub.schema(schema.id).get();
-      assert.isNotOk(got, "schema hadn't been deleted");
-    } catch (e) {
-      // This is expected.
-    }*/
+      // Because of caching delays, this can't be deterministic without a big wait.
+      /*try {
+        const got = await pubsub.schema(schema.id).get();
+        assert.isNotOk(got, "schema hadn't been deleted");
+      } catch (e) {
+        // This is expected.
+      }*/
+    } finally {
+      await cleanMine(id);
+    }
   });
 
   it('should get a schema', async () => {
-    const schema = await createSchema('proto');
+    const id = 'get-schema';
+    try {
+      const schema = await createSchema(id, 'proto');
 
-    const output = execSync(`${commandFor('getSchema')} ${schema.id}`);
-    assert.include(output, schema.id);
-    assert.include(output, 'info:');
-    assert.include(output, 'PROTO');
+      const output = execSync(`${commandFor('getSchema')} ${schema.id}`);
+      assert.include(output, schema.id);
+      assert.include(output, 'info:');
+      assert.include(output, 'PROTO');
+    } finally {
+      await cleanMine(id);
+    }
   });
 
   it('should listen for avro records', async () => {
-    const schema = await createSchema('avro');
-    const topic = await createTopicWithSchema(schema.id, Encodings.Json);
-    const sub = await createSub(topic.name);
+    const id = 'listen-avro';
+    try {
+      const schema = await createSchema(id, 'avro');
+      const topic = await createTopicWithSchema(id, schema.id, Encodings.Json);
+      const sub = await createSub(id, topic.name);
 
-    topic.publish(
-      Buffer.from(
-        JSON.stringify({
-          name: 'Alberta',
-          post_abbr: 'AB',
-        })
-      )
-    );
-    await topic.flush();
+      topic.publish(
+        Buffer.from(
+          JSON.stringify({
+            name: 'Alberta',
+            post_abbr: 'AB',
+          })
+        )
+      );
+      await topic.flush();
 
-    const output = execSync(
-      `${commandFor('listenForAvroRecords')} ${sub.name} 3`
-    );
-    assert.include(output, 'Received message');
-    assert.include(output, 'Alberta');
-    assert.include(output, 'AB');
+      const output = execSync(
+        `${commandFor('listenForAvroRecords')} ${sub.name} 3`
+      );
+      assert.include(output, 'Received message');
+      assert.include(output, 'Alberta');
+      assert.include(output, 'AB');
+    } finally {
+      await cleanMine(id);
+    }
   });
 
   it('should listen for protobuf messages', async () => {
-    const schema = await createSchema('proto');
-    const topic = await createTopicWithSchema(schema.id, Encodings.Json);
-    const sub = await createSub(topic.name);
+    const id = 'listen-proto';
+    try {
+      const schema = await createSchema(id, 'proto');
+      const topic = await createTopicWithSchema(id, schema.id, Encodings.Json);
+      const sub = await createSub(id, topic.name);
 
-    topic.publish(
-      Buffer.from(
-        JSON.stringify({
-          name: 'Quebec',
-          post_abbr: 'QC',
-        })
-      )
-    );
-    await topic.flush();
+      topic.publish(
+        Buffer.from(
+          JSON.stringify({
+            name: 'Quebec',
+            post_abbr: 'QC',
+          })
+        )
+      );
+      await topic.flush();
 
-    const output = execSync(
-      `${commandFor('listenForProtobufMessages')} ${sub.name} 3`
-    );
-    assert.include(output, 'Received message');
-    assert.include(output, 'Quebec');
-    assert.include(output, 'QC');
+      const output = execSync(
+        `${commandFor('listenForProtobufMessages')} ${sub.name} 3`
+      );
+      assert.include(output, 'Received message');
+      assert.include(output, 'Quebec');
+      assert.include(output, 'QC');
+    } finally {
+      await cleanMine(id);
+    }
   });
 
   it('should list schemas', async () => {
-    const schema = await createSchema('avro');
-    const output = execSync(`${commandFor('listSchemas')}`);
-    assert.include(output, schema.id);
+    const id = 'list';
+    try {
+      const schema = await createSchema(id, 'avro');
+      const output = execSync(`${commandFor('listSchemas')}`);
+      assert.include(output, schema.id);
+    } finally {
+      await cleanMine(id);
+    }
   });
 
   it('should publish avro records', async () => {
-    const schema = await createSchema('avro');
-    const topic = await createTopicWithSchema(schema.id, Encodings.Binary);
-    const sub = await createSub(topic.name);
-    const deferred = defer();
-    sub.on('message', deferred.resolve);
-    sub.on('error', deferred.reject);
+    const id = 'pub-avro';
+    try {
+      const schema = await createSchema(id, 'avro');
+      const topic = await createTopicWithSchema(
+        id,
+        schema.id,
+        Encodings.Binary
+      );
+      const sub = await createSub(id, topic.name);
+      const deferred = defer();
+      sub.on('message', deferred.resolve);
+      sub.on('error', deferred.reject);
 
-    const output = execSync(
-      `${commandFor('publishAvroRecords')} ${topic.name}`
-    );
-    assert.include(output, 'published.');
+      const output = execSync(
+        `${commandFor('publishAvroRecords')} ${topic.name}`
+      );
+      assert.include(output, 'published.');
 
-    const result = (await deferred.promise) as Message;
-    assert.include(result.data.toString(), 'Ontario');
+      const result = (await deferred.promise) as Message;
+      assert.include(result.data.toString(), 'Ontario');
 
-    sub.close();
+      sub.close();
+    } finally {
+      await cleanMine(id);
+    }
   });
 
   it('should publish protobuf messages', async () => {
-    const schema = await createSchema('proto');
-    const topic = await createTopicWithSchema(schema.id, Encodings.Binary);
-    const sub = await createSub(topic.name);
-    const deferred = defer();
-    sub.on('message', deferred.resolve);
-    sub.on('error', deferred.reject);
+    const id = 'pub-proto';
+    try {
+      const schema = await createSchema(id, 'proto');
+      const topic = await createTopicWithSchema(
+        id,
+        schema.id,
+        Encodings.Binary
+      );
+      const sub = await createSub(id, topic.name);
+      const deferred = defer();
+      sub.on('message', deferred.resolve);
+      sub.on('error', deferred.reject);
 
-    const output = execSync(
-      `${commandFor('publishProtobufMessages')} ${topic.name}`
-    );
-    assert.include(output, 'published.');
+      const output = execSync(
+        `${commandFor('publishProtobufMessages')} ${topic.name}`
+      );
+      assert.include(output, 'published.');
 
-    const result = (await deferred.promise) as Message;
-    assert.include(result.data.toString(), 'Ontario');
+      const result = (await deferred.promise) as Message;
+      assert.include(result.data.toString(), 'Ontario');
 
-    sub.close();
+      sub.close();
+    } finally {
+      await cleanMine(id);
+    }
   });
 });
